third_party/blink/web_tests/shapedetection/detection-support.html

<!DOCTYPE html>
<link rel="help" href="https://wicg.github.io/shape-detection-api/#dom-barcodedetector-getsupportedformats">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/services/shape_detection/public/mojom/barcodedetection.mojom.js"></script>
<script src="file:///gen/services/shape_detection/public/mojom/barcodedetection_provider.mojom.js"></script>
<script src="resources/mock-barcodedetection.js"></script>
<script>
'use strict';

promise_test(() => {
  return new Promise((resolve, reject) => {
    var formats = BarcodeDetector.getSupportedFormats();
    resolve(formats);
  }).then(result => {
    assert_equals(result.length, 3, 'Number of supported formats');
    assert_equals(result[0], 'aztec', 'format 1');
    assert_equals(result[1], 'data_matrix', 'format 2');
    assert_equals(result[2], 'qr_code', 'format 3');
  });
}, 'get supported barcode formats');

</script>
